requirements.txt
contourpy==1.3.2
cycler==0.12.1
fonttools==4.62.1
kiwisolver==1.5.0
matplotlib==3.10.8
numpy==2.2.6
packaging==26.0
pandas==2.3.3
pillow==12.1.1
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-mnist==0.7
pytz==2026.1.post1
six==1.17.0
tzdata==2025.3
